+++ frontend/backend/src/data-source.ts
@@ -2,6 +2,7 @@ import "reflect-metadata"
 import { DataSource } from "typeorm"
 import { Media } from "./entity/Media"
 import { Customer } from "./entity/Customer"
+import { Borrowing } from "./entity/Borrowing"
 
 export const AppDataSource = new DataSource({
     type: "mysql",
@@ -12,7 +13,7 @@ export const AppDataSource = new DataSource({
     database: "videokolcsonzo",
     synchronize: true,
     logging: false,
-    entities: [Media, Customer],
+    entities: [Media, Customer, Borrowing],
     migrations: [],
     subscribers: [],
 })