+++ frontend/backend/src/entity/Borrowing.ts
@@ -17,18 +17,18 @@ export class Borrowing implements BorrowingDTO {
   @JoinColumn({ name: "azonosito" })
   customer: Customer;
 
-  @Column()
+  @Column({ type: 'date', nullable: false })
   kolcsonzes_datuma: Date;
 
-  @Column({ nullable: true })
+  @Column({ type: 'date', nullable: true })
   visszahozas_datuma: Date | null;
 
   sorszam: number;
   azonosito: number;
 
-  constructor() {
+/*   constructor() {
     this.sorszam = this.media ? this.media.sorszam : 0;  // Ha van media, hozzuk el a sorszamot
     this.azonosito = this.customer ? this.customer.azonosito : 0;  // Ha van customer, hozzuk el az azonosito-t
-  }
+  } */
 }