+++ frontend/src/app/services/borrowing.service.ts
@@ -1,0 +1,16 @@
+import { inject, Injectable } from '@angular/core';
+import { HttpClient, HttpParams } from '@angular/common/http';
+import { BorrowingDTO } from '../../../models';
+
+@Injectable({ 
+    providedIn: 'root' 
+})
+
+export class BorrowingService {
+  http = inject(HttpClient);
+
+    getCustomerBorrowings(azonosito: number){
+      return this.http.get<BorrowingDTO[]>('/api/getCustomerBorrowings/' + azonosito);
+    }
+
+}